Move adapter implementations to new folder & modules. AdapterLayer -> BottleneckLayer.
