+++ src/adapters/models/albert/mixin_albert.py
@@ -3,10 +3,10 @@ from typing import Callable, Iterable, Tuple
 import torch.nn as nn
 
 from ...composition import adjust_tensors_for_parallel_
-from ...layer import AdapterLayer
-from ...lora import Linear as LoRALinear
+from ...methods.bottleneck import BottleneckLayer
+from ...methods.lora import Linear as LoRALinear
+from ...methods.prefix_tuning import PrefixTuningShim
 from ...model_mixin import EmbeddingAdaptersMixin, InvertibleAdaptersMixin, ModelBaseAdaptersMixin
-from ...prefix_tuning import PrefixTuningShim
 
 
 class AlbertAttentionAdaptersMixin:
@@ -18,7 +18,7 @@ class AlbertAttentionAdaptersMixin:
         self.key = LoRALinear.wrap(self.key, "selfattn", model_config, adapters_config, attn_key="k")
         self.value = LoRALinear.wrap(self.value, "selfattn", model_config, adapters_config, attn_key="v")
 
-        self.attention_adapters = AdapterLayer("mh_adapter")
+        self.attention_adapters = BottleneckLayer("mh_adapter")
 
         self.prefix_tuning = PrefixTuningShim(
             self.location_key + "_prefix" if self.location_key else None, model_config, adapters_config
@@ -36,7 +36,7 @@ class AlbertEncoderLayerAdaptersMixin:
         # Set location keys for prefix tuning
         self.location_key = "output_adapter"
 
-        self.output_adapters = AdapterLayer("output_adapter")
+        self.output_adapters = BottleneckLayer("output_adapter")
 
         self.attention.location_key = "self"
