+++ src/adapters/models/beit/mixin_beit.py
@@ -2,10 +2,10 @@ from typing import Iterable, Tuple
 
 import torch.nn as nn
 
-from ...layer import AdapterLayer
-from ...lora import Linear as LoRALinear
+from ...methods.bottleneck import BottleneckLayer
+from ...methods.lora import Linear as LoRALinear
+from ...methods.prefix_tuning import PrefixTuningShim
 from ...model_mixin import ModelBaseAdaptersMixin
-from ...prefix_tuning import PrefixTuningShim
 
 
 class BeitSelfAttentionAdaptersMixin:
@@ -38,8 +38,8 @@ class BeitLayerAdaptersMixin:
     """Adds adapters to the BeitLayer module."""
 
     def init_adapters(self, model_config, adapters_config):
-        self.attention_adapters = AdapterLayer("mh_adapter")
-        self.output_adapters = AdapterLayer("output_adapter")
+        self.attention_adapters = BottleneckLayer("mh_adapter")
+        self.output_adapters = BottleneckLayer("output_adapter")
 
 
 class BeitModelAdaptersMixin(ModelBaseAdaptersMixin):
