+++ src/adapters/models/vit/mixin_vit.py
@@ -2,10 +2,10 @@ from typing import Iterable, Tuple
 
 import torch.nn as nn
 
-from ...layer import AdapterLayer
-from ...lora import Linear as LoRALinear
+from ...methods.bottleneck import BottleneckLayer
+from ...methods.lora import Linear as LoRALinear
+from ...methods.prefix_tuning import PrefixTuningShim
 from ...model_mixin import ModelBaseAdaptersMixin
-from ...prefix_tuning import PrefixTuningShim
 
 
 class ViTSelfAttentionAdaptersMixin:
@@ -32,7 +32,7 @@ class ViTOutputAdaptersMixin:
     """Adds adapters to the ViTOutput module."""
 
     def init_adapters(self, model_config, adapters_config):
-        self.output_adapters = AdapterLayer("output_adapter")
+        self.output_adapters = BottleneckLayer("output_adapter")
 
         # Wrap layers for LoRA
         self.dense = LoRALinear.wrap(self.dense, "output", model_config, adapters_config)
@@ -43,7 +43,7 @@ class ViTLayerAdaptersMixin:
     """Adds adapters to the ViTSelfOutput module."""
 
     def init_adapters(self, model_config, adapters_config):
-        self.attention_adapters = AdapterLayer("mh_adapter")
+        self.attention_adapters = BottleneckLayer("mh_adapter")
 
 
 class ViTModelAdaptersMixin(ModelBaseAdaptersMixin):
